minor code restructure
