+++ src/mgkdb/support/mgk_file_handling.py
@@ -138,6 +138,75 @@ def f_check_required_files(global_vars, fldr, suffix, sim_type):
     
     return files_exist
 
+def f_user_input_metadata(database):
+    '''
+    Create a dictonary of user inputs for metadata
+    Used as keyword arguments to construct metadata dictionary
+    '''
+
+    user_ip = {} 
+    print("Filling metadata.")
+    skip_metadata = input("To skip entering Metadata, please enter 0. press any other key to continue\n")
+    if skip_metadata=='0': 
+        return user_ip
+
+    print("Please provide input for metadata. Press Enter to skip that entry.\n")
+    confidence = input('What is your confidence (1-10) for the run? Press ENTER to use default value -1.0\n')
+    if len(confidence):
+        confidence = float(confidence)
+    else:
+        confidence = -1.0
+        print("Using default confidence -1.\n")
+
+    user_ip['confidence']= confidence 
+
+    keywords = input('Any keywords to categorize this run. Press Enter to skip\n')
+    user_ip['keywords'] = keywords
+    
+    comments = input('Any comments for data in this folder?Press Enter to skip.\n')
+    user_ip['comments'] = comments
+
+    linked_id_strg = input('Do you want to link this run to another existing run in MGKDB. Press Enter to skip\n')
+    user_ip['linked_ID'] = f_get_linked_oid(database, linked_id_strg) if linked_id_strg else None
+
+    archive = input('Is there a location where the data is archived? Press Enter to skip.\n')
+    user_ip['archive_loc'] = archive
+
+    restart = input('Is this run a restart starting from a different run? For yes -> Y . no -> any other key \n')
+
+    user_ip['restart'] = (restart=='Y')
+
+    if restart=='Y':
+        user_ip['restart_timestep'] = int(input('What was the timestep of the previous run used to start this run?\n'))
+        initial_run_info = input('Has the initial run been uploaded to this database. For yes -> Y .\n')
+        run_oid = ObjectId(input('Please enter the ObjectID for that run\n')) if initial_run_info == 'Y' else None
+        if f_check_id_exists(database, run_oid):
+            user_ip['initial_run_oid'] =  run_oid 
+        else:
+            print(f"Entered object id {run_oid} doesn't exist\n")
+    
+    expt = input('Name of actual or hypothetical experiment? Eg: diiid, iter, sparc, etc. Press Enter to skip.\n')
+    user_ip['expt'] = expt
+
+    scenario_id = input('Scenario ID : shot ID or time or runID? Eg: 129913.1500ms . Press Enter to skip.\n')
+    user_ip['scenario_runid'] = scenario_id
+
+    git_hash = input('Do you have git-hash to store?Press Enter to skip.\n')
+    user_ip['git_hash'] = git_hash
+
+    platform = input('Platform on which this was run? Eg: perlmutter, summit, engaging, pc . Press Enter to skip.\n')
+    user_ip['platform'] = platform
+
+    exec_date = input('Execution date?Press Enter to skip.\n')
+    user_ip['ex_date'] = exec_date
+
+    workflow = input('Workflow type? Eg: portals, smarts, standalone, etc. Press Enter to skip.\n')
+    user_ip['workflow_type'] = workflow
+
+    print("Publication information should be uploaded with a separate script")
+
+    return user_ip
+
 def f_set_metadata(user=None,out_dir=None,suffix=None,keywords=None,confidence=-1,comments='Uploaded with default settings.',time_upload=None,\
                    last_update=None, linked_ID=None, expt=None, scenario_runid=None, linear=None, quasiLinear=None, has_1dflux = None, sim_type=None,\
                    git_hash=None, platform=None, ex_date=None, workflow_type=None, archive_loc=None, restart=False, restart_timestep=0, initial_run_oid=None):
